% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_data_files.R
\name{abs_data_files}
\alias{abs_data_files}
\title{Extract data form abs time series sheets}
\usage{
abs_data_files(file)
}
\arguments{
\item{file}{Name and path of input file}
}
\value{
list of data frames
}
\description{
This function only work's with data in \verb{Data 1, Data 2, ...} sheets.This is a helper function to \code{clean_abs_raw}
}
