% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_indexed_files.R
\name{abs_indexed_files}
\alias{abs_indexed_files}
\title{Extract data form abs time series sheets}
\usage{
abs_indexed_files(file)
}
\arguments{
\item{file}{Name and path of input file}
}
\value{
List of data frames
}
\description{
This function only work's with files containing \code{Index} sheet. This is a helper function to \code{clean_abs_raw}
}
