% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_tables_files.R
\name{abs_tables_files}
\alias{abs_tables_files}
\title{Extract data form abs time series sheets}
\usage{
abs_tables_files(file)
}
\arguments{
\item{file}{Name and path of input file}
}
\value{
list of data frames
}
\description{
This function only work's with data in \verb{Table 1, Table 2, ...} sheets. This is a helper function to \code{clean_abs_raw}
}
