% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_rates.R
\name{basic_rate}
\alias{basic_rate}
\title{Basic Rate}
\usage{
basic_rate(num, den, factor)
}
\arguments{
\item{num}{A number or a vector}

\item{den}{A number or a vector}

\item{factor}{A number}
}
\value{
Numeric
}
\description{
This function is for internal use to calculate some rates.
}
\keyword{internal}
