% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_change.R
\name{calc_change}
\alias{calc_change}
\alias{calc_change_per}
\title{Calculates the change}
\usage{
calc_change(column, len, dt)

calc_change_per(column, len, dt)
}
\arguments{
\item{column}{The column to use for calculating the change.}

\item{len}{The number of lags to use.}

\item{dt}{The date column to order the column values by.}
}
\value{
A vector containing the change values.
}
\description{
\code{calc_change} calculates the level change of a column in a dataframe over a specified amount of time,
\code{calc_change_per} calculates the percentage change of a column in a dataframe over a specified amount of time.
}
\examples{
library(dplyr)

lfs_state \%>\%
group_by(date) \%>\%
summarise(emp = sum(emp), .groups = 'drop') \%>\%
mutate(chng = calc_change(emp, 1, date))
}
