% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_abs_raw.R
\name{clean_abs_raw}
\alias{clean_abs_raw}
\title{Clean's ABS timeseries sheets}
\usage{
clean_abs_raw(file)
}
\arguments{
\item{file}{Name and path of input file}
}
\value{
clean data frame
}
\description{
This function tidy up ABS excel files and returns a clean data frame.
For this function to work the data should be a timeseries and saved in sheets called either Data or Table
}
\examples{
\dontrun{
# Read ABS file with data saved in Data 1 sheet
df <- clean_abs_raw("./path_to/EQ08.xlsx")
}
}
