% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_df.R
\name{corr_df}
\alias{corr_df}
\title{Correlation dataframe with p-values}
\usage{
corr_df(data)
}
\arguments{
\item{data}{numeric data frame}
}
\value{
correlation data frame with p-values in parenthesis
}
\description{
Correlation dataframe with p-values
}
\examples{

corr_df(datasets::iris[,-5])

}
