% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{example1}
\alias{example1}
\title{Example of data of accidentes in a company for a year.}
\format{
A data frame with 5 variables: \code{month}, \code{staff},
\code{hours.worked}, \code{work.accidents}, \code{accidents.on.your.way.home}
\describe{
\item{month}{Month by number}
\item{staff}{Number of staff}
\item{hours.worked}{The total of hours worked for all the workers on the month}
\item{work.accidents}{Number of accidents in the place of work}
\item{accidents.on.your.way.home}{Number of accidents on the workers' on their way home or on their way their place of work from their home}
}
}
\usage{
example1
}
\description{
Example of data of accidentes in a company for a year.
}
\keyword{datasets}
