% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.sfa.R
\name{extract.sfa}
\alias{extract.sfa}
\title{Title}
\usage{
extract.sfa(
  model,
  include.loglik = TRUE,
  include.eff = TRUE,
  include.nobs = TRUE,
  ...
)
}
\arguments{
\item{model}{model object}

\item{include.loglik}{GOF log-likelihood}

\item{include.eff}{GOF Mean efficiency}

\item{include.nobs}{GOF number of observations}

\item{...}{extra setting for sfa}
}
\value{
list of extract
}
\description{
Title
}
