% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_rates.R
\name{frequency_rate}
\alias{frequency_rate}
\title{Frequency rate}
\usage{
frequency_rate(hours, acci, factor = 10^6)
}
\arguments{
\item{hours}{Number of hours worked by workers in the reference group during
the reference period, or a vector to sum.}

\item{acci}{Number of new cases of occupational injury during the reference period,
or a vector to sum.}

\item{factor}{Multiplicative factor to adjust the rate (default is 10^6).}
}
\value{
Numeric value of the frequency rate
}
\description{
A frequency rate indicate the number of new cases of injury
occurring in relation to the amount of time during which workers in the reference group
were "exposed to the risk" of being involved in an occupational accident.
}
\examples{
# Example with one observation
frequency_rate(hours = 800000, acci = 80)
}
