% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_join_dates.R
\name{fuzzy_join_dates}
\alias{fuzzy_join_dates}
\title{Fuzzy join for dates}
\usage{
fuzzy_join_dates(extraction_date, release_date, n = 1, len_hist = 18)
}
\arguments{
\item{extraction_date}{The extraction_date column from a dataframe as date vector.}

\item{release_date}{The release_date column from a dataframe as date vector.}

\item{n}{The number of prior periods to find. Defaults to 1.}

\item{len_hist}{The length of history to use in months. Ensures that the matching release dates are not too old. Defaults to 18 months.}
}
\value{
A dataframe that enables fuzzy date join.
}
\description{
Finds the n release dates immediately prior to the extraction date, release dates must be prior to extraction dates .
}
\examples{
library(dplyr)
#data with release dates
df <- lfs_state |>
     group_by(anzsco1_code, date) |>
     summarise(emp = sum(emp), .groups = 'drop')

# extraction dates
e_dt <- as.Date(c("2016-08-09", "2021-08-10"))
# look at 4 prior quarters
n_qtr <- 4
join_dates <- fuzzy_join_dates(
e_dt,
df$date,
n = 3 * n_qtr,
len_hist = (n_qtr + 1) * 12
)
}
