% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_rates.R
\name{incidence_rate}
\alias{incidence_rate}
\title{Incidence rate}
\usage{
incidence_rate(workers, acci, factor = 10^3)
}
\arguments{
\item{workers}{Total number of workers in the reference group during the
reference period, or a vector to sum.}

\item{acci}{Number of new cases of occupational injury during the reference period,
or a vector to sum.}

\item{factor}{Multiplicative factor to adjust the rate (default is 10^3).}
}
\value{
Numeric value of the incidence rate
}
\description{
A frequency rate indicate the number of new cases of injury
occurring in relation to the amount of workers in the reference group.
}
\examples{
incidence_rate(workers = 800, acci = 80)
}
