% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsa_palettes.R
\name{jsa_palettes}
\alias{jsa_palettes}
\title{JSA colour palettes}
\usage{
jsa_palettes(
  name,
  n,
  all_palettes = jsa_colours,
  type = c("discrete", "continuous")
)
}
\arguments{
\item{name}{name of the colour palette we want to use}

\item{n}{how many colours from it we want to use}

\item{all_palettes}{list of colour palettes we want to extract our choice from}

\item{type}{discrete or continuous colour palette}
}
\value{
colour palette
}
\description{
JSA colour palettes
}
\examples{
library(ggplot2)

df <- data.frame(
x = c("A", "B", "C", "D"),
y = 1:4
)

g <- ggplot(
 data = df,
 mapping = aes(x = x, y = y)
 ) +
theme_minimal() +
 theme(
   legend.position = c(0.05, 0.95),
   legend.justification = c(0, 1),
   legend.title = element_blank(),
   axis.title = element_blank()
 )



}
