% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_state.R
\docType{data}
\name{lfs_state}
\alias{lfs_state}
\title{Labour Force Employment.}
\format{
A data frame with 9,728 rows and 5 variables:
\describe{
\item{state_name}{Australian states}
\item{anzsco1_code}{occupation major group of main job (ANZSCO) code}
\item{anzsco1_name}{occupation major group of main job (ANZSCO) name}
\item{date}{date}
\item{emp}{employments in thousands}
}
}
\source{
\url{https://www.abs.gov.au/statistics/labour/employment-and-unemployment/labour-force-australia-detailed/latest-release#labour-force-status}
}
\usage{
lfs_state
}
\description{
Sample of Australian labour force by state and occupation.
}
\keyword{datasets}
