% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_lines.R
\name{limits}
\alias{limits}
\title{Limits}
\usage{
limits(hours, fr, conf.level = 0.9)
}
\arguments{
\item{hours}{Number of hours worked by workers in the reference group during
the reference period, or a vector to sum.}

\item{fr}{Frequency rate expected}

\item{conf.level}{Confidence level of the interval.}
}
\value{
A list containing the following components:
\item{lim_inf}{The inferior limit.}
\item{lim_sup}{The superior limit.}
}
\description{
This function is internal to calculate the superior and inferior limits for
the frequency rate.
}
\keyword{internal}
