% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_rates.R
\name{safety_rate}
\alias{safety_rate}
\title{Safety Rate}
\usage{
safety_rate(workers, acci, hours, factor = 10^5)
}
\arguments{
\item{workers}{Total number of workers in the reference group during the
reference period, or a vector to sum.}

\item{acci}{Number of new cases of occupational injury during the reference period,
or a vector to sum.}

\item{hours}{Number of hours worked by workers in the reference group during
the reference period, or a vector to sum.}

\item{factor}{Multiplicative factor to adjust the rate (default is 10^5).}
}
\value{
Numeric value of the days lost rate
}
\description{
Safety Rate
}
\examples{
safety_rate(workers = 800, acci = 80, hours = 800000)
}
