% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_series.R
\name{separate_series}
\alias{separate_series}
\title{Separate ABS data items}
\usage{
separate_series(
  data,
  series_name = "series",
  column_names = NULL,
  remove_totals = FALSE,
  remove_nas = FALSE
)
}
\arguments{
\item{data}{data frame}

\item{series_name}{column series to expand}

\item{column_names}{Optional, names of new column series}

\item{remove_totals}{Optional, remove totals}

\item{remove_nas}{Optional, remove missing values}
}
\description{
Separate a character column into multiple columns
}
\examples{
\dontrun{
# Read ABS file with data saved in Data 1 sheet
df <- clean_abs_raw("./path_to/6202001.xlsx")
df |>
separate_series("data_item_description")
}
}
