% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_rates.R
\name{severity_rate}
\alias{severity_rate}
\title{Severity Rate}
\usage{
severity_rate(hours, days, factor = 10^3)
}
\arguments{
\item{hours}{Number of hours worked by workers in the reference group during
the reference period or a vector to sum.}

\item{days}{Number of days lost as a result of new cases of occupational
injury during the reference period, or a vector to sum. In case of permanent
incapacity each country have a scale to know how many days have to add for
this kind of workers.}

\item{factor}{Multiplicative factor to adjust the rate (default is 10^3).}
}
\value{
Numeric value of the severity rate
}
\description{
A severity rates mesure the time lost due to injures in relation to the total
amount of time owrked.
}
\examples{
severity_rate(hours = 800000, days = 300)
}
