% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_data_files.R
\name{tidy_abs_data}
\alias{tidy_abs_data}
\title{Tidy list into a data frame
This is a helper function to \code{clean_abs_raw}}
\usage{
tidy_abs_data(lst)
}
\arguments{
\item{lst}{list object from \code{abs_data_files}}
}
\value{
Clean data frame
}
\description{
Tidy list into a data frame
This is a helper function to \code{clean_abs_raw}
}
