% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_indexed_files.R
\name{tidy_abs_indexed}
\alias{tidy_abs_indexed}
\title{Tidy list into a data frame}
\usage{
tidy_abs_indexed(lst)
}
\arguments{
\item{lst}{object from \code{abs_indexed_files}}
}
\value{
clean data frame
}
\description{
Note this function takes \code{abs_indexed_files} object as an input.This is a helper function to \code{clean_abs_raw}
}
