% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_tables_files.R
\name{tidy_abs_tables}
\alias{tidy_abs_tables}
\title{Tidy list into a data frame}
\usage{
tidy_abs_tables(lst)
}
\arguments{
\item{lst}{list object from \code{abs_tables_files}}
}
\value{
Clean data frame
}
\description{
This is a helper function to \code{clean_abs_raw}
}
