% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Comparing models}
\usage{
compare_models(models, model_names = NULL)
}
\arguments{
\item{models}{A statistical model object or a list of statistical model objects. Lists of models can be specified as
models = list(model.1, model.2, ...). Different object types can also be mixed.}

\item{model_names}{A character vector of labels for the model names. By default,
the models are named "mod_1", "mod_2", etc. Specifying model_names = c("Model name 1", "Model name 2") etc. overrides the default behavior.}
}
\value{
A dataframe of compared models for easy model comparison
}
\description{
Comparing models
}
\examples{
library(plm)
library(dplyr)

data(worldprod, package = 'jsar')
countries <- c("Australia", "singapore", "china", "uk")

# comparison with same model (lm) for each country
lm_models <- worldprod |>
  filter(country_name \%in\% countries) |>
  group_by(country_name) |>
  group_map(~ lm(log_gdp ~ log_capital + log_labour , data = .x))

compare_models(lm_models, model_names = paste0('lm_',countries))

# or with piping
worldprod |>
  filter(country_name \%in\% countries) |>
  group_by(country_name) |>
  group_map(~ lm(log_gdp ~ log_capital + log_labour , data = .x)) |>
  compare_models( model_names = paste0('lm_',countries))

lm_model <- lm(log_gdp ~ log_capital + log_labour, data = worldprod)
plm_model <- plm::plm(log_gdp ~ log_capital + log_labour ,
                      model = 'within', index = c('country_name', 'year'),
                      data = worldprod)

compare_models(list(lm_model, plm_model),model_names = c('Pool', 'Panel'))

}
