% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Comparing models}
\usage{
compare_models(models, model_names = NULL)
}
\arguments{
\item{models}{an object containing the results returned by a model fitting function. Supported models are 'lm' & 'sfa', put multiple models in list}

\item{model_names}{vector of model names as string}
}
\value{
dataframe of compared models and list for individual models
}
\description{
Comparing models
}
\examples{

mod1 <- lm(Volume ~ Girth + Height, data = datasets::trees)
mod2 <- lm(Volume ~ Girth , data = datasets::trees)
# Comparing models
compare_models(list(mod1, mod2))$compare
compare_models(mod2)
}
