% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_rates.R
\name{frequency_rate}
\alias{frequency_rate}
\title{Frequency rate}
\usage{
frequency_rate(hours, acci, factor = 10^6)
}
\arguments{
\item{hours}{Number total of Hours or vector}

\item{acci}{Number total of accidents or vector}

\item{factor}{Multipliavate factor}
}
\value{
numeric
}
\description{
A frequency rate indicate the number of new cases of injury
occurring in relation to the amount of time during which workers in the reference group
were "exposed to the risk" of being involved in an occupational accident.
}
\examples{
frequency_rate(hours = 800000, acci = 80, factor = 10^6)
}
